Moved sub-classes of "ToolStripComboBox" and "ToolStripProgressBar" classes to their own files
